feat: CLI user management

+++ MC Controls/iam/adapter-out-persistence/src/main/kotlin/dev/lynne/mc_controls/iam/adapter_out_persistence/repository/CredentialJpaRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface CredentialJpaRepository : JpaRepository<CredentialJpaEntity, Long>
+interface CredentialJpaRepository : JpaRepository<CredentialJpaEntity, Int>
